RFC physically correct projectile start velocity
The current teeworlds projectile velocity is calculated exclusively by
the tunables *Speed. This patch adds the tee's velocity to the
projectile. This has an big impact on the game, so it's a RFC merge
request. However this will give the player more opportunities to shoot
projectiles, but it is more difficult to handle because the projectiles don't
fly in the aimed direction.

Signed-off-by: Markus Pargmann <mpargmann@allfex.org>